Orientation is a float

+++ zoloto/coords.py
@@ -5,7 +5,7 @@ from coordinates import spaced_coordinate
 Coordinates = spaced_coordinate("Coordinates", "xy")
 
 Orientation = NamedTuple(
-    "Orientation", [("rot_x", int), ("rot_y", int), ("rot_z", int)]
+    "Orientation", [("rot_x", float), ("rot_y", float), ("rot_z", float)]
 )
 
 ThreeDCoordinates = spaced_coordinate("ThreeDCoordinates", "xyz")
